+++ src/facade/mod.rs
@@ -2,3 +2,4 @@ pub mod list_handler;
 pub mod help_handler;
 pub mod add_handler;
 pub mod misc_handler;
+pub mod clear_handler;