+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Jotdown
 
+## version 0.5.01
+- 10/28/2023
+- 3 more features pending to make first release!
+- Completed clear command `jd clear --done`
+- Completed renumber command `jd renumber`
+- POC: Send push notification to Mobile using `ntfy.sh` curl command!!
+- Unit testing and refactoring
+
 ## version 0.5.00
 - 10/15/2023
 - Removed unused library crates - rustc-serializer and regex